bias-audit-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
scikit-learn==1.8.0
xgboost==3.2.0
shap==0.51.0
joblib==1.5.3
anthropic==0.86.0
supabase==2.9.0
python-dotenv==1.2.2
numpy==2.4.3
pandas==3.0.1
pydantic==2.12.5
